+++ tests/test-678rf64.rb
@@ -1,0 +1,5 @@
+#!/usr/bin/ruby -w
+
+# T_678rf64
+describe "mkvmerge / reading WAV of type RF64"
+test_merge "data/wav/v.rf64"